refactor!: blockPositions for shard methods are shardRelative
* rename and better document methods taking blockPositions
* refactoring
* update tests

+++ src/main/java/org/janelia/saalfeldlab/n5/shard/InMemoryShard.java
@@ -35,7 +35,7 @@ public class InMemoryShard<T> extends AbstractShard<T> {
 	}
 
 	@Override
-	public DataBlock<T> getBlock(long... blockGridPosition) {
+	public DataBlock<T> getBlock(int... blockGridPosition) {
 
 		return blocks.get(Position.wrap(blockGridPosition));
 	}
